clean up audiocaps
