+++ mteb/tasks/Audio/Any2AnyRetrieval/AudioCaps.py
@@ -4,11 +4,11 @@ from mteb.abstasks.Audio.AbsTaskAny2AnyRetrieval import AbsTaskAny2AnyRetrieval
 from mteb.abstasks.TaskMetadata import TaskMetadata
 
 
-class AudioCapsRetrieval(AbsTaskAny2AnyRetrieval):
+class AudioCapsA2TRetrieval(AbsTaskAny2AnyRetrieval):
     metadata = TaskMetadata(
-        name="AudioCapsRetrieval",
-        description="Measuring the ability to retrieve the groundtruth answers to reasoning task queries on ARC-Challenge.",
-        reference="https://allenai.org/data/arc",
+        name="AudioCapsA2TRetrieval",
+        description="Natural language description for any kind of audio in the wild.",
+        reference="https://audiocaps.github.io/",
         dataset={
             "path": "AudioLLMs/audiocaps_test",
             "revision": "fb42aac15212cbddd723fbbf04b6071b60a9f8fe",
@@ -27,18 +27,12 @@ class AudioCapsRetrieval(AbsTaskAny2AnyRetrieval):
         dialect=[],
         sample_creation="found",
         bibtex_citation=r"""
-@article{clark2018think,
-  author = {Clark, Peter and Cowhey, Isaac and Etzioni, Oren and Khot, Tushar and Sabharwal, Ashish and Schoenick, Carissa and Tafjord, Oyvind},
-  journal = {arXiv preprint arXiv:1803.05457},
-  title = {Think you have solved question answering? try arc, the ai2 reasoning challenge},
-  year = {2018},
-}
-
-@article{xiao2024rar,
-  author = {Xiao, Chenghao and Hudson, G Thomas and Moubayed, Noura Al},
-  journal = {arXiv preprint arXiv:2404.06347},
-  title = {RAR-b: Reasoning as Retrieval Benchmark},
-  year = {2024},
+@inproceedings{kim2019audiocaps,
+  author = {Kim, Chris Dongjoo and Kim, Byeongchang and Lee, Hyunmin and Kim, Gunhee},
+  booktitle = {Proceedings of the 2019 Conference of the North American Chapter of the Association for Computational Linguistics: Human Language Technologies, Volume 1 (Long and Short Papers)},
+  pages = {119--132},
+  title = {Audiocaps: Generating captions for audios in the wild},
+  year = {2019},
 }
 """,
         # prompt={"query": "Retrieve the answer to the question."},
@@ -47,6 +41,59 @@ class AudioCapsRetrieval(AbsTaskAny2AnyRetrieval):
     audio_column_name: str = "context"
     text_column_name: str = "answer"
     id_column_name: str = "audiocap_id"
+
+    def dataset_transform(self):
+        for split in self.dataset.keys():
+            self.dataset[split] = self.dataset[split]
+
+            self.dataset[split] = self.dataset[split].add_column(
+                "audiocap_id",
+                [f"audiocap_{i}" for i in range(len(self.dataset[split]))],
+            )
+
+
+class AudioCapsT2ARetrieval(AbsTaskAny2AnyRetrieval):
+    """This reverses the normal AudioCaps dataset. Instead of audio queries to text corpus, this tests text queries to audio corpus."""
+
+    metadata = TaskMetadata(
+        name="AudioCapsT2ARetrieval",
+        description="Natural language description for any kind of audio in the wild.",
+        reference="https://audiocaps.github.io/",
+        dataset={
+            "path": "AudioLLMs/audiocaps_test",
+            "revision": "fb42aac15212cbddd723fbbf04b6071b60a9f8fe",
+        },
+        type="Retrieval",
+        category="t2a",
+        modalities=["text", "audio"],
+        eval_splits=["test"],
+        eval_langs=["eng-Latn"],
+        main_score="ndcg_at_10",
+        date=("2018-01-01", "2018-12-31"),
+        domains=["Encyclopaedic", "Written"],
+        task_subtypes=["Reasoning as Retrieval"],
+        license="cc-by-sa-4.0",
+        annotations_creators="derived",
+        dialect=[],
+        sample_creation="found",
+        bibtex_citation=r"""
+@inproceedings{kim2019audiocaps,
+  author = {Kim, Chris Dongjoo and Kim, Byeongchang and Lee, Hyunmin and Kim, Gunhee},
+  booktitle = {Proceedings of the 2019 Conference of the North American Chapter of the Association for Computational Linguistics: Human Language Technologies, Volume 1 (Long and Short Papers)},
+  pages = {119--132},
+  title = {Audiocaps: Generating captions for audios in the wild},
+  year = {2019},
+}
+""",
+    )
+
+    audio_column_name: str = "context"
+    text_column_name: str = "answer"
+    id_column_name: str = "audiocap_id"
+
+    # reverse the default modalities
+    default_query_modality: str = "text"  # queries will be text
+    default_corpus_modality: str = "audio"  # corpus will be audio
 
     def dataset_transform(self):
         for split in self.dataset.keys():
